Move capturePhysicsTrajectories & animateBlastWithGSAP created by Jake to utils/animationHelpers.js
